reactor: Rename
