+++ module/data/src/main/kotlin/io/github/siyual_park/data/repository/mongo/MongoRepositoryBuilder.kt
@@ -1,8 +1,8 @@
 package io.github.siyual_park.data.repository.mongo
 
 import com.google.common.cache.CacheBuilder
-import io.github.siyual_park.data.cache.InMemoryNestedStorage
 import io.github.siyual_park.data.cache.InMemoryStorage
+import io.github.siyual_park.data.cache.NestedStorage
 import io.github.siyual_park.data.cache.Pool
 import io.github.siyual_park.data.cache.TransactionalStorage
 import io.github.siyual_park.data.repository.Extractor
@@ -44,7 +44,7 @@ class MongoRepositoryBuilder<T : Any, ID : Any>(
         return if (cacheBuilder != null) {
             val idExtractor = createIdExtractor(current)
             val storage = TransactionalStorage(
-                InMemoryNestedStorage(Pool { InMemoryStorage(cacheBuilder, idExtractor) }, idExtractor)
+                NestedStorage(Pool { InMemoryStorage(cacheBuilder, idExtractor) }, idExtractor)
             )
 
             CachedMongoRepository(current, storage, idExtractor)
